requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-chroma
chromadb
pypdf
python-dotenv
tiktoken
sentence-transformers
googlesearch-python
requests
beautifulsoup4
torch
fpdf2
datetime
